pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ['tastychef', "alekseenkolesha@yandex.ru>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django = "<5"
python-dotenv = "^0.20.0"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"
django-bootstrap5 = "^23.3"
dj-database-url = "^2.0.0"
psycopg2-binary = "^2.9.3"
django-filter = "^22.1.0"
rollbar = "^0.16.3"
tzdata = "*"
pytz = "*"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
coverage = "^6.4.2"
black = "^24.4.2"
isort = "*"
coveralls = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
